SDALGCP2: Fast Spatially Discrete Approximation to Log-Gaussian Cox Processes

A faster, modernised successor to the SDALGCP package for fitting a spatially discrete approximation to a log-Gaussian Cox process (SDA-LGCP) to spatially aggregated disease counts. Performance-critical steps are implemented in C++ via RcppArmadillo. See DESIGN.md in the package source for the full design and statistical roadmap.

See also

Author

Maintainer: Olatunji Johnson teejaystatistics@gmail.com

Authors:

Emanuele Giorgi

Peter Diggle